udp-music: fix fec_group_size config parsing
popl's Value<uint8_t> reads "4" as ASCII char 0x34 = 52, not integer 4.
Widen fec_group_size to size_t; cast back to uint8_t on use. Default
N=4 now survives a round-trip through the config file.

+++ server/server.cpp
@@ -458,7 +458,8 @@ void Server::start()
         if (settings_.udp_stream.enabled)
         {
             udpAudioServer_ = std::make_unique<UdpAudioServer>(io_context_, settings_.udp_stream.bind_to_address,
-                                                               settings_.udp_stream.port, settings_.udp_stream.fec_group_size);
+                                                               settings_.udp_stream.port,
+                                                               static_cast<uint8_t>(settings_.udp_stream.fec_group_size));
         }
         streamManager_ = std::make_unique<StreamManager>(this, io_context_, settings_);
